+++ src/equicordplugins/tidalControls/TidalPlayer.tsx
