+++ src/equicordplugins/tidalControls/hoverOnly.css
