+++ src/equicordplugins/tidalControls/index.tsx
