+++ src/equicordplugins/tidalControls/tidalStyles.css
